Page fixes (Scaling & Positioning)
I've updated the elements of the page to properly fit, I now have a section to hold the background images.

Restructured some of the CSS code to properly fit the screen, here's no longer issues with the page creating a scroll bar. Using a fixed position is the solution to any issue such as that.