Image attempt (Background)
I've attempted to add a background image to the page, though I've encountered a sizing issue, currently present in the code.

To fix this, I'll have to find the right values for the bottom-text box (gray area), or find a way to smoothly transition between the image and the bottom area (gray area) without a gap.