Adjustments (Scaling Issues)
Adjusting the page to fit the image better, I've encountered yet another issue with my design. Changing the pages size seems to cause issues due to the images fixed size, I'll either have to set a minimum and maximum size, find a new solution, or allow the page to have a scroll bar.

Further testing and feedback needed, this will be the last iteration for today. For my next session, I'll make sure to get new images from the Stakeholder that may fit better to the website, a size such as 1080p could work well, though I need to figure out a way to fix the scaling issues.

To add, I've removed the gradient on the bottom area (gray area), and instead will attempt to have an overlay to the image, giving the image the gradient instead. I'm not sure how well this will work though.