Navigation Bar Updates
Today I've managed to work on the navigation bar more, adding the placeholder for the buttons to outline where they will be. From here, updating the UI to have buttons is easy, as I've worked out the values for sizing, and positioning. To add, I've included the website's title in the navigation bar, as chosen in the high fidelity.

While I didn't get the chance to get new photos, I've made the most of my time and done all I can in the time allocated. Hopefully in the next session I'll have the updates photos for the background, though that's entirely dependent on the Stakeholder providing them.

From here, I need to update the background. As discussed in the previous update, I'm yet to find a way to solve the sizing issues. Though, I've found good examples by looking through other websites.