Background Fixes Update
Using W3Schools, I've managed to fix the background image issue, where it would either get cut off, size incorrectly, or move other parts of the website.

Now that this issue is fixed, I can work to finish this page.

To add, I've adjusted the navigation bar to have the correct buttons, all that's left is to add an underline to each button, and link them to other pages.