Some major changes
Added a footer bar to later hold a sign in system. And fixed some sizing issues I had noticed.

To add, I've started working on the other pages, linking up the navigation bars and separating some of the CSS.

I've created a javascript file to manage the slideshow in the gallery page.