Quick fixes
I've updates the slideshow to allow the viewer to tap it to navigate between images. While there's no back button, the Stakeholder is happy with the simplicity of the design.

While I encountered some issues with the timed loop not resetting when clicked, I've managed to get it to a functional state.