Major Updates
Working with my Stakeholder, I've managed to work on the website to a standard the Stakeholder considers complete. Though some final touches will be made, such as adding more photos, text, and adjusting some of the colours.

To add, I've adjusted some of the pages, and reformatted them to PHP, removing some of the issues I had with it changing tab when the user submits their email to the newsletters.

Finishing this update, I've managed to fully incorporate a database using PHP and HTML.