Final Update
With the database fully implimented, the Stakeholder is asking for the website to be finished. With no time left, we've decided to end it here, leaving the about page to be filled with text by them later once they come up with something to put on the page.

The Stakeholder is happy with the results presented, saying that's it's exactly what they need.

In this update, I've made some minor changes to some of the code, adjusting some background colors and the JavaScript code for the Gallery page.

To add, I've included a SVG file, adding some 'noise' to the background colors in most of the pages. This makes the website more aesthetically pleasing to look at, rather than solid colors.